Build from storage account

+++ .jenkins/docker/base/build.sh
@@ -13,7 +13,7 @@ IMAGE_TAG="latest"
 usage() {
     echo "Usage: $0 -m <string> -o <string> [OPTIONS..]" 1>&2
     echo "Create a base Docker image for Mystikos" 1>&2
-    echo "  -m     Mystikos release version (Example: 0.5.0)"
+    echo "  -m     Mystikos release version (Example: 0.5.0) or path to Mystikos tar.gz"
     echo "           See https://github.com/deislabs/mystikos/releases for release versions"
     echo "  -o     Open Enclave Base Docker Image tag (Examples: latest, SGX-2.15.100)"
     echo "           See https://hub.docker.com/repository/docker/oeciteam/openenclave-base-ubuntu-18.04 for release versions"
@@ -40,8 +40,6 @@ while getopts "hm:o:u::t::" option; do
     esac
 done
 
-MYST_TARBALL="Ubuntu-${UBUNTU_VERSION//.}_mystikos-${MYSTIKOS_RELEASE_VERSION}-x86_64.tar.gz"
-
 # Catch extra parameters
 shift "$((OPTIND-1))"
 if [[ ! -z "${1}" ]]; then
@@ -66,16 +64,22 @@ case "${UBUNTU_VERSION}" in
            ;;
 esac
 
-# Download Mystikos release
-if [[ -f "${BUILD_DIR}/${MYST_TARBALL}" ]]; then
-    rm "${BUILD_DIR}/${MYST_TARBALL}"
+if [[ "${MYSTIKOS_RELEASE_VERSION}" =~ [0-9]+\\.[0-9]+\\.[0-9]+ ]]; then
+    # Download Mystikos release
+    MYST_TARBALL="Ubuntu-${UBUNTU_VERSION//.}_mystikos-${MYSTIKOS_RELEASE_VERSION}-x86_64.tar.gz"
+    if [[ -f "${BUILD_DIR}/${MYST_TARBALL}" ]]; then
+        rm "${BUILD_DIR}/${MYST_TARBALL}"
+    fi
+    wget \
+    --directory-prefix="${BUILD_DIR}" \
+    --no-verbose \
+    --tries=3 \
+    --waitretry=3 \
+    https://github.com/deislabs/mystikos/releases/download/v${MYSTIKOS_RELEASE_VERSION}/${MYST_TARBALL}
+else
+    echo "Mystikos release version does not match [0-9]+.[0-9]+.[0-9]+ ... assuming local tarball path."
+    MYST_TARBALL="${MYSTIKOS_RELEASE_VERSION}"
 fi
-wget \
-  --directory-prefix="${BUILD_DIR}" \
-  --no-verbose \
-  --tries=3 \
-  --waitretry=3 \
-  https://github.com/deislabs/mystikos/releases/download/v${MYSTIKOS_RELEASE_VERSION}/${MYST_TARBALL}
 
 # Build Docker image
 set -x
